refactor: separate journey steps into components
